Created individual component for LoginForm
